Updates to Build Plan and Gen Cases.
Build Plan - Updated build_plan to ignore lines with Exec_with_creds.
Previous script was removing the "_with_creds" and replacing it with a
comment.

Gen Cases - Remove the need to look for yyyy.mm-patch.json file. Instead
use a generic bundle-patch.json. This file can be use for each patch
cycle since the subject line will describe the active bundle we are
working on.