Updates to build_plan and new template.
Added new variable to pull the first host from a large host list. This
is useful for the versioncheck command when I only need one server to
search. 

Added razor_batch template to do servers in batches and gsize variable
to determine the concurrency of how many to do in parallel.